added player respawn

+++ core/src/com/mygdx/game/entities/Flame.java
@@ -113,8 +113,14 @@ public class Flame extends Entity implements PhysicsObject {
     @Override
     public void handleBeginContact(PhysicsObject psycho2, GameWorld world) {
         if (psycho2 instanceof Walker) return;
-        psycho2.setFlagForDelete(true);
-        world.getEntityManager().removeEntity((Entity)psycho2);
+        if (psycho2 instanceof Player) {
+            Player player = (Player)psycho2;
+            player.timeout();
+
+        } else {
+            psycho2.setFlagForDelete(true);
+            world.getEntityManager().removeEntity((Entity)psycho2);
+        }
         erupt();
     }
